added About Us page and moved flexbox section with cards from Home Page to About Us page
