+++ frontend/src/Routes.js
@@ -2,6 +2,7 @@ import { BrowserRouter as Router, Route } from 'react-router-dom';
 import ReviewsPage from './containers/ReviewsPage/ReviewsPage'
 import ProfilePage from './containers/ProfilePage/ProfilePage';
 import HomePage from './containers/HomePage/HomePage';
+import AboutUsPage from './containers/AboutUsPage/AboutUsPage';
 
 export default function Routes() {
     return (
@@ -12,6 +13,7 @@ export default function Routes() {
             */}
             <Route path="/reviews" exact component={ReviewsPage}></Route>
             <Route path="/profile" exact component={ProfilePage}></Route>
+            <Route path="/aboutus" exact component={AboutUsPage}></Route>
             <Route path="/" exact component={HomePage}></Route>
 
         </Router>
