+++ frontend/src/containers/AboutUsPage/AboutUsPage.css
@@ -1,0 +1,18 @@
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+}
+
+html {
+    width:100%;
+    height:100%;
+}
+
+.aboutUsPage-container {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    width:100%;
+    justify-content: space-evenly;
+}
