+++ frontend/src/containers/HomePage/HomePage.css
@@ -8,11 +8,3 @@ html {
     width:100%;
     height:100%;
 }
-
-.homePage-container {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    width:100%;
-    justify-content: space-evenly;
-}
